Move metadata overlap checks to its own module
